sustainability_data.py
import requests
from bs4 import BeautifulSoup
import os
from dotenv import load_dotenv

load_dotenv()

class SustainabilityDataCollector:
    def __init__(self):
        self.good_on_you_api_key = os.getenv('GOOD_ON_YOU_API_KEY')
        self.base_url = "https://api.goodonyou.eco/api/v1"

    def get_brand_rating(self, brand_name):
        """Get sustainability rating from Good On You API"""
        headers = {'Authorization': f'Bearer {self.good_on_you_api_key}'}
        response = requests.get(
            f"{self.base_url}/brands/{brand_name}",
            headers=headers
        )
        return response.json() if response.status_code == 200 else None

    def scrape_fashion_revolution_data(self):
        """Scrape latest sustainability data from Fashion Revolution"""
        url = "https://www.fashionrevolution.org/transparency-index/"
        response = requests.get(url)
        soup = BeautifulSoup(response.content, 'html.parser')
        # Add specific scraping logic based on website structure
        return {}

    def get_combined_sustainability_data(self, brand_name):
        """Combine data from multiple sources"""
        data = {
            'good_on_you': self.get_brand_rating(brand_name),
            'fashion_revolution': self.scrape_fashion_revolution_data()
        }
        return data